Move DbUtils implemetation to separate package. Move helper functions out of interface
