+++ testing/db-testkit/src/main/kotlin/net/corda/db/testkit/dbutilsimpl/DbUtilsHelper.kt
@@ -1,9 +1,8 @@
-package net.corda.db.testkit
+package net.corda.db.testkit.dbutilsimpl
 
 import com.typesafe.config.Config
 import net.corda.db.core.CloseableDataSource
 import net.corda.orm.EntityManagerConfiguration
-import java.sql.Connection
 
 /**
  * An interface defining common database utility methods for managing database connections and configurations.
@@ -43,21 +42,4 @@ interface DbUtilsHelper {
         dbPassword: String? = null,
         schemaName: String? = null
     ): Config
-
-    fun getPropertyNonBlank(key: String, defaultValue: String): String {
-        val value = System.getProperty(key)
-        return if (value.isNullOrBlank()) {
-            defaultValue
-        } else {
-            value
-        }
-    }
-
-    fun Connection.createSchema(schemaName: String?) {
-        requireNotNull(schemaName)
-        this.use { conn ->
-            conn.prepareStatement("CREATE SCHEMA IF NOT EXISTS $schemaName;").execute()
-            conn.commit()
-        }
-    }
-}
+ }
