+++ testing/db-testkit/src/main/kotlin/net/corda/db/testkit/dbutilsimpl/PostgresHelper.kt
@@ -1,4 +1,4 @@
-package net.corda.db.testkit
+package net.corda.db.testkit.dbutilsimpl
 
 class PostgresHelper : AbstractDBHelper() {
     companion object {
